requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.4,<5.0
python-dotenv>=1.0,<2.0
openpyxl>=3.1,<4.0
reportlab>=4.0,<5.0
